rtptc: fr: Keep 'Lunatic' wording
Commit 00be06cfe0127 changed 'Lunatique' to 'Aliéné' for a rtptc card.
This 'Aliéné' change is something specific to the revised edition.
There are no 'return to' printing with this wording change.
Making this change is also inconsistent with the English version which
hasn't changed.
This commit restores the previous wording for the card to use the same
terms as the printed card, and avoid confusion for ArkhamCards French
users.